+++ example/poisson/parthenon_app_inputs.cpp
@@ -30,8 +30,8 @@ using namespace parthenon;
 
 namespace poisson_example {
 
-void ProblemGenerator(MeshBlock *pmb, ParameterInput *pin) {
-  auto &data = pmb->meshblock_data.Get();
+void ProblemGenerator(Mesh *pm, ParameterInput *pin, MeshData<Real> *md) {
+  auto pmb = md->GetBlockData(0)->GetBlockPointer();
 
   Real x0 = pin->GetOrAddReal("poisson", "x0", 0.0);
   Real y0 = pin->GetOrAddReal("poisson", "y0", 0.0);
@@ -43,16 +43,17 @@ void ProblemGenerator(MeshBlock *pmb, ParameterInput *pin) {
   IndexRange jb = cellbounds.GetBoundsJ(IndexDomain::entire);
   IndexRange kb = cellbounds.GetBoundsK(IndexDomain::entire);
 
-  auto coords = pmb->coords;
   PackIndexMap imap;
   const std::vector<std::string> vars({"density", "potential"});
-  const auto &q = data->PackVariables(vars, imap);
+  const auto &q_bpack = md->PackVariables(vars, imap);
   const int irho = imap["density"].first;
   const int iphi = imap["potential"].first;
 
   pmb->par_for(
-      "Poisson::ProblemGenerator", kb.s, kb.e, jb.s, jb.e, ib.s, ib.e,
-      KOKKOS_LAMBDA(const int k, const int j, const int i) {
+      "Poisson::ProblemGenerator", 0, q_bpack.GetDim(5) - 1, kb.s, kb.e, jb.s, jb.e, ib.s,
+      ib.e, KOKKOS_LAMBDA(const int b, const int k, const int j, const int i) {
+        const auto &coords = q_bpack.GetCoords(b);
+        auto &q = q_bpack(b);
         Real dist2 = std::pow(coords.x1v(i) - x0, 2) + std::pow(coords.x2v(j) - y0, 2) +
                      std::pow(coords.x3v(k) - z0, 2);
         if (dist2 < radius * radius) {