+++ src/mesh/meshblock.hpp
@@ -259,8 +259,6 @@ class MeshBlock : public std::enable_shared_from_this<MeshBlock> {
     pbval->SearchAndSetNeighbors(tree, ranklist, nslist);
   }
 
-  void ResetToIC() { ProblemGenerator(nullptr, nullptr); }
-
   // inform MeshBlock which arrays contained in member Field, Particles,
   // ... etc. classes are the "primary" representations of a quantity. when registered,
   // that data are used for (1) load balancing (2) (future) dumping to restart file
@@ -419,8 +417,7 @@ class MeshBlock : public std::enable_shared_from_this<MeshBlock> {
 
   // defined in either the prob file or default_pgen.cpp in ../pgen/
   static void ProblemGeneratorDefault(MeshBlock *pmb, ParameterInput *pin);
-  std::function<void(MeshBlock *, ParameterInput *)> ProblemGenerator =
-      &ProblemGeneratorDefault;
+  std::function<void(MeshBlock *, ParameterInput *)> ProblemGenerator = nullptr;
   static pMeshBlockApplicationData_t
   InitApplicationMeshBlockDataDefault(MeshBlock *, ParameterInput *pin);
   std::function<pMeshBlockApplicationData_t(MeshBlock *, ParameterInput *)>